decode_short_ids: move to common.
We want to use it in devtools/decodemsg.
