+++ gossipd/Makefile
@@ -48,6 +48,7 @@ GOSSIPD_COMMON_OBJS :=				\
 	common/cryptomsg.o			\
 	common/daemon.o				\
 	common/daemon_conn.o			\
+	common/decode_short_channel_ids.o	\
 	common/dev_disconnect.o			\
 	common/features.o			\
 	common/gen_status_wire.o		\
